Chapter 4 - Assignment (completed)
Hi professor,
The assignment for Ch.4 is now completed.
The main folder contains the files for Q2, and the folder "Q2_Program" the files for Q2.
In addition, the document "Languages and Paradigms Chapter 4 Project1_Santi_v1.pdf" includes my responses with details for both Q1 and Q2.
In "README.txt" this is explained in detail.
If you have any question, just let me know!
Thank you,